+++ models/Course.js
@@ -1,0 +1,8 @@
+const mongoose = require('mongoose');
+
+const Course = mongoose.model('Course', {
+  title: {type: String, required: true},
+  description: String
+})
+
+module.exports = Course
